Move all actions from index into separate files AuthActions, UserActions, DogActions, and PlaydateActions
